Fix banner dismiss option persistence
