+++ client/settings/payment-methods/index.js
@@ -1,23 +1,14 @@
 /* global wc_stripe_settings_params */
 import { __ } from '@wordpress/i18n';
-import React, { useContext, useState } from 'react';
+import React from 'react';
 import { ExternalLink } from '@wordpress/components';
 import SettingsSection from '../settings-section';
 import PaymentRequestSection from '../payment-request-section';
 import GeneralSettingsSection from '../general-settings-section';
 import LoadableSettingsSection from '../loadable-settings-section';
 import DisplayOrderCustomizationNotice from '../display-order-customization-notice';
-import {
-	BNPL_PROMOTION_BANNER,
-	NEW_CHECKOUT_EXPERIENCE_BANNER,
-	OC_PROMOTION_BANNER,
-} from 'wcstripe/settings/payment-settings/constants';
+import { NEW_CHECKOUT_EXPERIENCE_BANNER } from 'wcstripe/settings/payment-settings/constants';
 import PromotionalBanner from 'wcstripe/settings/payment-settings/promotional-banner';
-import UpeToggleContext from 'wcstripe/settings/upe-toggle/context';
-import { useAccount } from 'wcstripe/data/account';
-import { useEnabledPaymentMethodIds } from 'wcstripe/data';
-import { getPromotionalBannerType } from 'wcstripe/settings/payment-settings/promotional-banner/get-promotional-banner-type';
-import OCToggleContext from 'wcstripe/settings/oc-toggle/context';
 
 const PaymentMethodsDescription = () => {
 	return (
@@ -56,36 +47,14 @@ const PaymentRequestDescription = () => (
 	</>
 );
 
-const PaymentMethodsPanel = ( { onSaveChanges } ) => {
-	const { data } = useAccount();
-	const { isOCEnabled, setIsOCEnabled } = useContext( OCToggleContext );
-	const { isUpeEnabled, setIsUpeEnabled } = useContext( UpeToggleContext );
-	const [ enabledPaymentMethodIds ] = useEnabledPaymentMethodIds();
-	const promotionalBannerType = getPromotionalBannerType(
-		data,
-		isUpeEnabled,
-		isOCEnabled,
-		enabledPaymentMethodIds
-	);
-	let initialBannerState = false;
-	if (
-		promotionalBannerType === BNPL_PROMOTION_BANNER &&
-		// eslint-disable-next-line camelcase
-		wc_stripe_settings_params?.show_bnpl_promotional_banner === '1'
-	) {
-		initialBannerState = true;
-	}
-	if (
-		promotionalBannerType === OC_PROMOTION_BANNER &&
-		// eslint-disable-next-line camelcase
-		wc_stripe_settings_params?.show_oc_promotional_banner === '1'
-	) {
-		initialBannerState = true;
-	}
-	const [ showPromotionalBanner, setShowPromotionalBanner ] = useState(
-		initialBannerState
-	);
-
+const PaymentMethodsPanel = ( {
+	onSaveChanges,
+	setShowPromotionalBanner,
+	showPromotionalBanner,
+	promotionalBannerType,
+	setIsOCEnabled,
+	setIsUpeEnabled,
+} ) => {
 	return (
 		<>
 			{ showPromotionalBanner && (
